backend/python/requirements.txt
fastapi[standard]==0.113.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
cachetools==5.5.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2
aiosqlite==0.20.0